Clang bugprone (#1040)
* bugprone and inline lint ignore of use after move

* integer division

* access specifiers, avoiding float in loops

* cuda

* cuda

+++ test/integration/test_chapman_integration.cpp
@@ -87,7 +87,7 @@ TEST(ChapmanIntegration, CanBuildChapmanSystem)
   state.conditions_[0].temperature_ = 2;
   state.conditions_[0].pressure_ = 3;
 
-  for (double t{}; t < 100; ++t)
+  for (int t{}; t < 100; ++t)
   {
     state.custom_rate_parameters_[0] = photo_rates;
     solver.UpdateStateParameters(state);
